LP init faster, PCs of Partial paths
Tried to make the initialization of the linear program faster by using
preallocated matrices rather than sparse ones. Sparse matricies still need to be
used when solving the LP or else it is prohibitively slow. It is the conversion
from dense to sparse matrices that seems to be the slowest part. One possible
solution might be to store these matrices beforehand. You can then have a fixed
number of nodes per frame: if the analysis yields more, the weakest are
discarded and if the analysis yields less dummy nodes of astronomical cost are
added. Note that the calculation of the cost matrix is quick, it is the
constraint matrices that take time. The number of paths can also be changed
quickly.

Tried solving a number of LPs in pursuit of a different number of paths to see
how the final cost behaves. It seems there is a slight inflection point when
pursuing a high number of paths, but it is very slight. My reckoning is that the
initial peak picking algorithm is good for harmonic sounds (that's what it was
tried on) and so there are few spurious peaks. We should try with more spurious
peaks to see what happens with the inflection point.

The original partial organization algorithm and the partial tracking algorithms
are not really compatible. To find a new way to group the partials into
different sources, I tried computing principal components on the
frequency-modulation/frequency values collected into vectors, each vector
representing the values a particular partial takes on. The results look
promising and it is with this that I think we should group the partials, after
they have been connected using a linear program.