Solve grid of LPs
In ddm_lp_test_{3,4}.py we try solving the partial tracking problem on a grid in
the time-frequency plane provided by the STFT. This works and makes the problems
relatively quick to solve by reducing their dimensionality. What needs to be
done now is we need to observe the principal components of the extracted paths
to see if they can be separated nicely.

harm_sines.m was improved to make it actually work and you can now run a script
to synthesize sounds (see harm_sines_script.m)