Comparing different DDM dimensions
Comparing 3 frames 1 bin, 1 frame 3 bins and 3 frames 3 bins and observing how
these are sensitive to the presence of multiple components.

Also tried with a minimum 4-term blackman window, modified to be conitnuous at
the boundaries. This does seem to improve the estimation of at least the
amplitude modulation (other parameters have not yet been estimated).