FM and AM function compare on decaying sinusoids
We are trying to extract good partial trajectories from a mixture of decaying
sinusoids provided by a recording of real instruments. To do this we must solve
for the partial paths.

We are interested in solving long K-best paths problems with few nodes in each
frame, and for this we will try employing the Viterbi algorithm. We are working
on converting the Viterbi solution to a linear program solution.