Guitar partial track and AM investigation
Adjusted parameters to better track partials of an acoustic guitar mixture. This
involved:
- Thresholding the cost/length ratio of the partial tracks by an L shaped
  function (an exponential), allowing tracks of very short length to pass.
- Adjusting the width of the bands in which peaks are searched for the partial
  path
- Adjusting the sensitivity of the peak picker.

We then investigated the relationship between the frequency and the slope of the
partial, the length of the partial and the beginning amplitude of the partial.
So far we observe no correlation that clearly separates the mixture of guitar
sounds.

See the following scripts for partial tracking, they save the partial tracks using
the pickle module in Python.
test/partial_estimation_plot_3_3b.py
test/partial_estimation_plot_3_3c.py

The following scripts plot the partial tracks:
test/plot_pep_3_3b.py
test/plot_pep_3_3c.py