Cubic phase synthesis with DDM
Tried the cubic sythesis method described by McAulay and Quatieri. Then extended
it to incorporate estimated frequency and frequency slope. The results are good.
Now want to carry out the same on the amplitude, but need to find a way to
constrain the amplitudes at the measured points, as well as solve for error
minimizing parameters via LS. We do this via constrained least squares. See
test/constrained_ls_test.m for an example of how to do it (unfortunately there
are no built in routines in numpy it seems).